Homes/Contacts/api/serializers.py
from rest_framework import serializers
from Homes.Contacts.models import Enquiry


class EnquirySerializer(serializers.ModelSerializer):
    class Meta:
        model = Enquiry
        fields = '__all__'
