requirements.txt
xhtml2pdf
watchtower
amqp==2.2.1
billiard==3.5.0.3
boto3==1.4.6
botocore==1.6.8
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.32.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-jsonfield==1.0.1
django-registration==2.2
django-rest-auth==0.9.1
django-storages==1.6.5
djangorestframework==3.6.4
docutils==0.14
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
feedparser==5.2.1
geopy==1.11.0
gunicorn==19.7.1
humanize==0.5.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
kombu==4.1.0
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.13.1
oauthlib==2.0.2
olefile==0.44
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyfcm==1.4.2
Pygments==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
python-http-client==3.0.0
python3-openid==3.1.0
pytz==2017.2
raven==6.10.0
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
s3transfer==0.1.10
scikit-learn==0.19.0
scipy==0.19.1
sendgrid==5.3.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.4
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
django-map-widgets==0.1.9
django-dirtyfields==1.3.1
django-celerybeat-status==0.0.10
pycrypto==2.6.1
google-api-python-client==1.6.7
google-cloud-vision==0.31.1
google-cloud==0.33.1
django-mysql==3.2.0
razorpay==1.2.0